feat: decompose ModelGenerator into factories, service objects, and type mappings
Refactor the monolithic ModelGenerator (2,176 lines) into decomposed classes:
- FieldFactory: field class creation from Metaschema definitions
- AssemblyFactory: assembly class creation with XML/JSON mapping
- Utils: shared safe naming utilities
- FieldSerializer/FieldDeserializer: service object pipeline for SINGLETON_OR_ARRAY
- CollapsiblesCollapser: collapsible field grouping support

Additional improvements:
- Add missing Metaschema data types to TypeMapper (day-time-duration,
  uri-reference, email-address, ip-v4-address, ip-v6-address, base64)
- Create MarkupMultilineDatatype as distinct type from MarkupLineDatatype
- Add JSON_VALUE_KEY constants for data-type-based defaults
- Add UNWRAPPED XML field handling with delegate mappings
- Emit delegate: option in RubySourceEmitter
- Require moxml >= 0.1.15 for namespace_validation_mode= API
